requirements.txt
asgiref==3.8.1
Django
django-cors-headers
django-extensions==3.2.3
djangorestframework==3.15.2
djangorestframework_simplejwt
PyJWT==2.9.0
PyMySQL==1.1.1
sqlparse==0.5.3
typing_extensions==4.12.2
